Move processSearchKeyword into search util
